chatbotService/requirements.txt
Flask==3.0.3
google-generativeai==0.7.1
python-dotenv==1.0.1
Flask-Cors==4.0.1
redis==5.0.7